+++ src/ui/callgrind.rs
@@ -287,7 +287,7 @@ mod tests {
     use crate::ui::callgrind::*;
 
     // Build a test stackframe
-    fn f(i: u32) -> StackFrame {
+    fn f(i: usize) -> StackFrame {
         StackFrame {
             name: format!("func{}", i),
             relative_path: format!("file{}.rb", i),
