+++ src/ui/summary.rs
@@ -134,7 +134,7 @@ mod tests {
     use crate::ui::summary::*;
 
     // Build a test stackframe
-    fn f(i: u32) -> StackFrame {
+    fn f(i: usize) -> StackFrame {
         StackFrame {
             name: format!("func{}", i),
             relative_path: format!("file{}.rb", i),
